Run as user on mac
